DAOS-4898 obj: add D_CASSERT for task mem space usage
Hit many times of run-time error that task mem space overflow, this
patch add static assertion at compiling time to avoid that.
