+++ src/common/tse_internal.h
@@ -33,8 +33,7 @@
  * Author: Di Wang  <di.wang@intel.com>
  */
 
-/* NB: tse_task_private is TSE_PRIV_SIZE = 1016 bytes for now */
-#define TSE_TASK_ARG_LEN		888
+#include <daos/tse.h>
 
 struct tse_task_private {
 	struct tse_sched_private	*dtp_sched;
